Hibernate instead of jdbc

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/daos/SimpleDao.java
@@ -5,6 +5,6 @@ import java.util.Optional;
 
 public interface SimpleDao<T> {
     T save(T t);
-    Optional<T> findById(long id);
+    Optional<T> findById(int id);
     List<T> findAll();
 }
